Check for end of buffer before reading char
This allows us to not force a null character at the end of the buffer
array. I still kept the check for explicit null characters after reading
in case there are null characters embedded in the middle of the string
(for some reason), as SMS would terminate early in that case. This also
fixes encoded text in binary formats saving with the null character.