Add pane-manipulation classes
All classes are assigned a pane and transform the attached pane over
time in various ways:

- bloExPane moves, resizes, and fades a pane linearly.
- bloBoundPane moves and resizes a pane in a 3-point curve.
- bloBlendPane inherits the capabilities of bloBoundPane, but can also
linearly fade a pane between two textures. This functionality works only
if attached to a bloPicture (the other functions work on any pane as
normal).

bloCoord2D is a utility type for linearly walking a 2D vector from one
value to another over time.